sclang: Fix scaling of canvas painting
The QPainter::drawPixmap we were using does scaling. We need to use the version that clips, in case the thing we're painting extends beyond the surface of the canvas.
Also, fix two places where we demote float pixelRatio to int - this would have broken fractional pixelRatios.

+++ QtCollider/widgets/QcScopeShm.cpp
@@ -165,7 +165,7 @@ void QcScopeShm::paintEvent ( QPaintEvent * event )
     int chanCount = _shm->reader.channels();
     int maxFrames = _shm->reader.max_frames();
     QRect area (_pixmap.rect());
-    area.setSize(area.size() / (int)_pixmap.devicePixelRatio());
+    area.setSize(area.size() / _pixmap.devicePixelRatio());
     
     p.begin(&_pixmap);
